xhci: tegra: fix a compilation warning in fw_log_wait_empty_timeout
Input parameter to time_is_after_jiffies() has to be in "unsigned long"
type.

bug 1320242

Reviewed-on: http://git-master/r/253185
(cherry picked from commit 604612466faae0d70cf08a9d02d8a9bf5d7749b3)

Reviewed-on: http://git-master/r/273704
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>

+++ drivers/usb/host/xhci-tegra.c
@@ -819,7 +819,7 @@ static inline bool fw_log_available(struct tegra_xhci_hcd *tegra)
 static inline bool fw_log_wait_empty_timeout(struct tegra_xhci_hcd *tegra,
 		unsigned timeout)
 {
-	u32 target = jiffies + msecs_to_jiffies(timeout);
+	unsigned long target = jiffies + msecs_to_jiffies(timeout);
 	bool ret;
 
 	mutex_lock(&tegra->log.mutex);
